refactor(device): extract the device validation flow to DeviceValidationController
Removed the responsibility for orchestrating the device registration validation
start and invalidation flows from DeviceController.

- Created DeviceValidationController with start and invalidate methods
- Updated routes to point to the new controller
- Renamed and adjusted related test files to reflect the new controller structure
